app/requirements.txt
gunicorn
psycopg2-binary
django==5.1.2
setuptools
djangorestframework
